test(e2e): move tests inside tests/lib
