+++ package.json
@@ -28,7 +28,7 @@
     "build": "tsc",
     "test": "jest --testPathIgnorePatterns ./tests/lib/mongo",
     "test:cov": "jest --coverage",
-    "test:e2e": "NODE_ENV=test jest --config ./tests/e2e/jest-e2e.json --setupFiles dotenv/config",
+    "test:e2e": "NODE_ENV=test jest --config ./tests/lib/e2e/jest-e2e.json --setupFiles dotenv/config",
     "test-mongo-init": "ts-node ./tests/lib/mongo/init",
     "test-mongo": "jest ./tests/lib/mongo --testTimeout 30000 --runInBand",
     "clean": "ts-node ./bin/delete-objects",
